+++ src/main/java/com/example/demo/repositories/OfertaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfertaRepository extends CrudRepository<OfertaModel, Integer> {
+public interface OfertaRepository extends CrudRepository<OfertaModel, Long> {
 }
