+++ src/main/java/com/example/demo/repositories/ProfesorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfesorRepository extends CrudRepository<ProfesorModel, Integer> {
+public interface ProfesorRepository extends CrudRepository<ProfesorModel, Long> {
 }
