+++ src/main/java/com/example/demo/repositories/ProgramaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProgramaRepository extends CrudRepository<ProgramaModel, Integer> {
+public interface ProgramaRepository extends CrudRepository<ProgramaModel, Long> {
 }
